storage: clear key-values with outdated version
when storing key-values (by calling the `set`/`delete` functions), all key-values with a `txId` smaller than the minimum ID among the multiple ongoing transactions should be cleaned up incidentally. Since the involved key-values may not only exist in the current leaf page to be written to, it is necessary to continuously process the previous pages. Therefore, the `pageHeader` needs to additionally record information such as `prevPageId` to enable navigation between pages.